server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.5",
    "mcp>=1.13.1",
    "physics-tenpy>=1.0.6",
    "quspin>=1.0.0",
]
